refactor: move event config to event_processor.rs and rename fields
