+++ ampd/src/tests/config_template.toml
@@ -2,11 +2,11 @@ health_check_bind_addr = '0.0.0.0:3000'
 tm_jsonrpc = 'http://localhost:26657/'
 tm_grpc = 'tcp://localhost:9090'
 
-[event_processor_config]
-retry_timeout = '1s'
+[event_processor]
+retry_delay = '1s'
 retry_max_attempts = 3
 stream_timeout = '15s'
-buffer_cap = 100000
+stream_buffer_size = 100000
 
 [broadcast]
 chain_id = 'axelar-dojo-1'
